Move the PDF extractor initialization to the external services common place
